requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
pandas
tqdm
Pillow
scikit-learn
trimesh
scipy
pyrender
